refactor: 💡 Minor client example refactoring
