+++ examples/client/package.json
@@ -18,7 +18,8 @@
     "react-dom": "^18.2.0",
     "typescript": "^4.9.3",
     "vite": "^4.1.0",
-    "zod": "^3.21.4"
+    "zod": "^3.21.4",
+    "@windingtree/sdk": "^0.1.0-beta.3"
   },
   "eslintConfig": {
     "extends": [
