+++ examples/server/index.ts
@@ -1,6 +1,6 @@
 import { createServer, ServerOptions } from '../../src/index.js';
 import { memoryStorage } from '../../src/storage/index.js';
-import peerKey from '../../test/peerKey.json';
+import peerKey from '../../test/peerKey.json' assert { type: 'json' };
 import { createLogger } from '../../src/utils/logger.js';
 
 const logger = createLogger('ServerMain');
